[#7] feat: deleteUser 유저 확인에 rdb 조회 대신 세션 사용으로 변경

+++ src/main/java/com/ludensdomain/service/UserService.java
@@ -79,6 +79,8 @@ public class UserService {
      * @param password   사용자 패스워드
      */
     public void deleteUser(long id, String password) {
-        userMapper.deleteUser(id, password);
+        if (loginService.verifyUser(id)) {
+            userMapper.deleteUser(id, password);
+        }
     }
 }
